@@ -5,16 +5,14 @@
     "babel-runtime": "^6.23.0",
     "es6-promise": "^4.1.0",
     "lodash": "4.17.4",
-    "react": "15.4.2",
-    "react-dom": "15.4.2",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-proxy-loader": "^0.3.5",
-    "react-redux": "^5.0.3",
+    "react-redux": "^5.0.4",
     "react-router": "^3.0.2",
     "react-tap-event-plugin": "2.0.1",
     "redux": "3.6.0",
     "uuid": "3.0.1"
-  },
-  "devDependencies": {
-    "react-addons-test-utils": "^15.4.2"
   }
 }
