@@ -1,5 +1,5 @@
 import React     from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import NotFound  from './not_found';
 import Stub      from '../../../../specs_support/stub';
 
