@@ -1,4 +1,4 @@
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import _ from 'lodash';
 
 export default {
